static
The static directory is a repository of static web application resources containing images for the entire site. Within static, separate catalogs are allocated for each specialty, where images specific to the corresponding specialty are stored. The index_banners subdirectory stores the images used on the main page of the site.